Added some debugging, cut file in 3, redesign leaderboard
